[Refactor] LogViewController rename, timeSticks: [View] refactoring
