+++ Project_Timer/Global/Extension/Date+Extension.swift
@@ -1,0 +1,17 @@
+//
+//  Date+Extension.swift
+//  Project_Timer
+//
+//  Created by Kang Minsang on 2022/03/14.
+//  Copyright © 2022 FDEE. All rights reserved.
+//
+
+import Foundation
+
+extension Date {
+    var MDstyleString: String {
+        let dateFormatter = DateFormatter()
+        dateFormatter.dateFormat = "M/d"
+        return dateFormatter.string(from: self)
+    }
+}
