+++ Project_Timer/Stopwatch/StopwatchViewController.swift
@@ -391,7 +391,7 @@ extension StopwatchViewController {
     }
     
     func showLog() {
-        let setVC = storyboard?.instantiateViewController(withIdentifier: "GraphViewController2") as! GraphViewController2
+        let setVC = storyboard?.instantiateViewController(withIdentifier: "GraphViewController2") as! LogViewController
             setVC.logViewControllerDelegate = self
             present(setVC,animated: true,completion: nil)
     }
